+++ tests/test_standalone_stac_item.py
@@ -36,11 +36,11 @@ def test_standalone_stac_item_thredds_ncml():
     # FIXME: avoid hackish workarounds
     data = requests.get(thredds_ncml_url).text
     attrs = xncml.Dataset.from_text(data).to_cf_dict()
-    attrs["access_urls"] = {  # FIXME: all following should be automatically added, but they are not!
+    attrs["access_urls"] = {
         "HTTPServer": f"{thredds_url}/fileServer/{thredds_path}/{thredds_nc}",
         "OPENDAP": f"{thredds_url}/dodsC/{thredds_path}/{thredds_nc}",
-        "WCS": f"{thredds_url}/wcs/{thredds_path}/{thredds_nc}?service=WCS&version=1.0.0&request=GetCapabilities",
-        "WMS": f"{thredds_url}/wms/{thredds_path}/{thredds_nc}?service=WMS&version=1.3.0&request=GetCapabilities",
+        "WCS": f"{thredds_url}/wcs/{thredds_path}/{thredds_nc}",
+        "WMS": f"{thredds_url}/wms/{thredds_path}/{thredds_nc}",
         "NetcdfSubset": f"{thredds_url}/ncss/{thredds_path}/{thredds_nc}/dataset.html",
     }
     stac_item = CMIP6Helper(attrs, GeoJSONPolygon).stac_item()