+++ CHANGES.md
@@ -7,6 +7,13 @@
 * Add an `xscen` extension demonstrating how to add properties to a STAC Item.
 * Fix mismatch between CMIP6 schema URI given to `pystac` and the actual schema URI
 * Add ability to export data from a STAC catalog or API to files on disk.
+* Add support for THREDDS services added in version 5.
+* Fix bug where session details weren't being used to access THREDDS catalogs.
+* Remove `title` from THREDDS catalog links since the content was specific to Marble.
+* Change link `type` from THREDDS catalog link from `text/xml` to `application/xml` (STAC API doesn't support `text/xml` anymore).
+* Support THREDDS datasets that don't contain an `NCML` access url.
+* Support nested collections when loading STAC objects with the `DirectoryLoader`.
+* `DirectoryLoader` now supports loading STAC objects created by the `export` command.
 
 ## [0.7.0](https://github.com/crim-ca/stac-populator/tree/0.7.0) (2025-03-07)